Chunk 5: Split alternative-location scoring into per-location scoring, ranking, and summary augmentation
- Extract per-location scoring to per-location.ts
- Extract ranking rules to ranking.ts
- Extract summary augmentation to augmentation.ts
- Extract debug projection to debug.ts
- Create types.ts for shared types
- Reduce score-alternatives.ts from 376 to 89 lines (-287 lines)
- Preserve all ranking thresholds and close-contender behavior

Closes #113
